# doc/api.yaml
openapi: 3.0.0
info:
  title: 'Nest NodeJS 2021Q2'
  description: 'NestJS, TypeORM, PostgresDB'
  version: '1.0'
  contact: {}
...
servers: []
components:
  schemas:
...
    CreateTaskDto:
      type: object
      properties:
        title:
          type: string
        order:
          type: number
        description:
          type: string
        userId:
          type: string
        boardId:
          type: string
        columnId:
          type: string
      required:
        - title
        - order
        - description
        - userId
        - boardId
        - columnId
    ResponseTaskDto:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        order:
          type: number
        description:
          type: string
        userId:
          type: string
        boardId:
          type: string
        columnId:
          type: string
      required:
        - id
        - title
        - order
        - description
        - userId
        - boardId
        - columnId
    UpdateTaskDto:
      type: object
      properties: {}
paths:
...
  '/boards/{boardId}/tasks':
    post:
      operationId: TasksController_create
      summary: 'Create one Task 👻'
      parameters:
        -
          name: boardId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskDto'
      responses:
        '201':
          description: 'The found record'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseTaskDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      security:
        -
          bearer: []
    get:
      operationId: TasksController_findAllByBoardId
      summary: 'Retrieve many Tasks 👻'
      parameters:
        -
          name: boardId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: 'The found records'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResponseTaskDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      security:
        -
          bearer: []
  '/boards/{boardId}/tasks/{id}':
    get:
      operationId: TasksController_findOneByBoardIdAndTaskId
      summary: 'Retrieve one Task 👻'
      parameters:
        -
          name: boardId
          required: true
          in: path
          schema:
            type: string
        -
          name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: 'The found record'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseTaskDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      security:
        -
          bearer: []
    put:
      operationId: TasksController_updateByBoardIdAndTaskId
      summary: 'Update one Task 👻'
      parameters:
        -
          name: boardId
          required: true
          in: path
          schema:
            type: string
        -
          name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskDto'
      responses:
        '200':
          description: 'The updated record'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseTaskDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      security:
        -
          bearer: []
    delete:
      operationId: TasksController_removeByBoardIdAndTaskId
      summary: 'Delete one Task 👻'
      parameters:
        -
          name: boardId
          required: true
          in: path
          schema:
            type: string
        -
          name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: 'The deleted record'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseTaskDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      security:
        -
          bearer: []